[Parser] separate into multiple files
